+++ dev/xtdb_tutorial.clj
@@ -7,31 +7,31 @@
 
 ;; prep -------------------------------
 
-(defn make-id-positive [infile outfile]
-  (fs/write-lines
-   outfile
-   (for [line (fs/read-all-lines infile)]
-     (str/replace line #"-(\d{3})" #(str (second %))))))
+; (defn make-id-positive [infile outfile]
+;   (fs/write-lines
+;    outfile
+;    (for [line (fs/read-all-lines infile)]
+;      (str/replace line #"-(\d{3})" #(str (second %))))))
 
-(make-id-positive "resources/docs.edn" "doc/docs-positive.edn")
+; (make-id-positive "resources/docs.edn" "doc/docs-positive.edn")
 
-(def my-docs
-  (clojure.edn/read-string (slurp "doc/docs-positive.edn")))
+; (def my-docs
+;   (clojure.edn/read-string (slurp "doc/docs-positive.edn")))
 
-(def schema {:movie/cast {:db/cardinality :db.cardinality/many}})
+; (def schema {:movie/cast {:db/cardinality :db.cardinality/many}})
 
-(def conn (start {:schema schema :url "jdbc:sqlite:storage/tutorial.sqlite"}))
+; (def conn (start {:schema schema :url "jdbc:sqlite:storage/tutorial.sqlite"}))
 
-(transact! conn my-docs)
+; (transact! conn my-docs)
 
-(comment
-  (ds/gc)
-  (ds/stop)
-  (def conn nil)
-  (ds/conn?)
-  :rcf)
+; (comment
+;   (ds/gc)
+;   (ds/stop)
+;   (def conn nil)
+;   (ds/conn?)
+;   :rcf)
 
-(stop)
+; (stop)
 
 ;; retart ----------