+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 # 0.7.3-SNAPSHOT
 
-- log level
-- argument of start must be a map or
+- restore's argument is a map {:url "jdbc-url"} or a string "jdc-url"
+- argument of start must be a map {:schema {schema} :url "url"}
+- consistency of log level
 
 # 0.7.2 (2025-09-01)